+++ daangnMarket/src/main/java/org/sopt/daangnMarket/controller/ItemController.java
@@ -1,7 +1,7 @@
 package org.sopt.daangnMarket.controller;
 
 import lombok.RequiredArgsConstructor;
-import org.sopt.daangnMarket.dto.Item.request.ItemCreateDto;
+import org.sopt.daangnMarket.dto.request.item.ItemCreateDto;
 import org.sopt.daangnMarket.service.ItemService;
 import org.sopt.daangnMarket.util.ApiResponse;
 import org.sopt.daangnMarket.util.ApiUtils;
@@ -12,8 +12,6 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-
-import java.net.URI;
 
 @RestController
 @RequiredArgsConstructor
