+++ daangnMarket/src/main/java/org/sopt/daangnMarket/dto/request/member/MemberCreateDto.java
@@ -1,4 +1,4 @@
-package org.sopt.daangnMarket.dto.Member.request;
+package org.sopt.daangnMarket.dto.request.member;
 
 public record MemberCreateDto(
         String nickname,
