+++ daangnMarket/src/main/java/org/sopt/daangnMarket/service/MemberService.java
@@ -3,8 +3,8 @@ package org.sopt.daangnMarket.service;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.sopt.daangnMarket.domain.Member;
-import org.sopt.daangnMarket.dto.Member.request.MemberCreateDto;
-import org.sopt.daangnMarket.dto.Member.response.MemberFindDto;
+import org.sopt.daangnMarket.dto.request.member.MemberCreateDto;
+import org.sopt.daangnMarket.dto.response.member.MemberFindDto;
 import org.sopt.daangnMarket.exception.ApiErrorCode;
 import org.sopt.daangnMarket.exception.DuplicateMemberException;
 import org.sopt.daangnMarket.exception.NotFoundException;
